+++ backend/src/main/java/com/example/webduck/member/service/port/MemberRepository.java
@@ -13,4 +13,5 @@ public interface MemberRepository {
     Optional<Member> findByUsername(String username);
     boolean existsByUsername(String username);
 
+    boolean existsById(Long id);
 }
