+++ backend/src/test/java/com/example/webduck/mock/member/WithMockCustomUserSecurityContextFactory.java
@@ -1,7 +1,7 @@
-package com.example.webduck.member.customMock;
+package com.example.webduck.mock.member;
 
 import com.example.webduck.global.security.oauth.entity.SessionMember;
-import com.example.webduck.member.entity.Member;
+import com.example.webduck.member.domain.Member;
 import java.util.List;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
